refactor: 자동차를 전진시키는 기능을 RacingGame에 넣기
